WIP - wip updated second fragment to use coroutines. need to handle network failure o coroutine
