+++ app/src/main/java/com/example/studentandroidapp/SecondFragment.kt
@@ -9,6 +9,14 @@ import android.widget.Button
 import android.widget.TextView
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
+import com.example.studentandroidapp.models.Student
+import com.example.studentandroidapp.network.StudentRestApi
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+import retrofit2.HttpException
+import retrofit2.Response
 
 /**
  * A simple [Fragment] subclass as the second destination in the navigation.
@@ -33,6 +41,30 @@ class SecondFragment : Fragment() {
 
         view.findViewById<Button>(R.id.button_second).setOnClickListener {
             findNavController().navigate(R.id.action_SecondFragment_to_FirstFragment)
+        }
+
+        view.findViewById<Button>(R.id.button_get_student_id_one).setOnClickListener {
+            displayStudentName()
+        }
+    }
+
+    private fun displayStudentName() {
+        val service = StudentRestApi.createRetrofitService()
+        var response  : Response<Student>? = null
+        CoroutineScope(Dispatchers.IO).launch {
+            try {
+                response = service.getStudentById("1")
+            } catch (e: Throwable) {
+                print(e.message)
+            }
+            withContext(Dispatchers.Main) {
+                    if (response?.isSuccessful!!) {
+                        view?.findViewById<TextView>(R.id.textview_second)?.text =
+                            response!!.body().toString()
+                    } else {
+                        view?.findViewById<TextView>(R.id.textview_second)?.text=getString(R.string.network_error)
+                    }
+            }
         }
     }
 }
